# docs/swagger.yaml
basePath: /api
definitions:
...
  model.Page:
    properties:
      nextCursor:
        type: string
    type: object
  model.PostLikeResponse:
    properties:
      likeCount:
        type: integer
    type: object
...
  model.ServerCommentCreateRequest:
    properties:
      content:
        type: string
      parentId:
        type: string
    type: object
  model.ServerCommentListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.ServerCommentResponse'
        type: array
      page:
        $ref: '#/definitions/model.Page'
    type: object
  model.ServerCommentResponse:
    properties:
      authorId:
        type: string
      content:
        type: string
      createDatetime:
        type: string
      id:
        type: string
      parentId:
        type: string
      updateDatetime:
        type: string
    type: object
...
  model.ServerPostListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.ServerPostResponse'
        type: array
      page:
        $ref: '#/definitions/model.Page'
    type: object
  model.ServerPostResponse:
    properties:
      caption:
        type: string
      commentCount:
        type: integer
      createDatetime:
        type: string
      likeCount:
        type: integer
      ownerId:
        type: string
      ownerImageUrl:
        type: string
      ownerName:
        type: string
      postId:
        type: string
      postImageUrl:
        type: string
      updateDatetime:
        type: string
    type: object
...
  model.ValidationError:
    properties:
      code:
        type: string
      message:
        type: string
      param:
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  description: Virdan Project API Documentation
  title: Virdan API
  version: "1.0"
paths:
...
  /posts/{postId}:
    get:
      description: |
        ## Get Post Flow

        ### Overview
        Retrieves a single post by ID. Requires membership in the server where the post belongs.

        ### Auth
        Requires `Authorization` header with Bearer JWT access token.

        ### Business Logic
        1. Get `userId` from context
        2. Parse `postId` from URL path — must be a valid UUID
        3. Check if user is a member of the server where the post belongs (single query)
        4. Fetch post details with author info, image URL, comment count, and like count

        ### Error Cases
        - Invalid postId → `400` with "Invalid post id"
        - Not a member of the server → `400` with "You are not a member of this server"

        ### Flow
        ```
        Request → Auth Middleware → Parse PostId → Check Server Membership via Post (DB) → Get Post (DB) → Response
        ```
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Post UUID
        in: path
        name: postId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ServerPostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ValidationError'
      summary: Get a single post
      tags:
      - posts
  /posts/{postId}/comments:
    get:
      description: |
        ## Get Comments Flow

        ### Overview
        Retrieves a paginated list of comments for a specific post. Requires membership in the server where the post belongs. Uses cursor-based pagination.

        ### Auth
        Requires `Authorization` header with Bearer JWT access token.

        ### Business Logic
        1. Get `userId` from context
        2. Parse `postId` from URL path — must be a valid UUID
        3. Parse query parameters: `limit`, `cursor`
        4. Validate `limit` (>= 0 and <= max limit)
        5. Check if user is a member of the server where the post belongs
        6. If cursor provided, decode and unmarshal
        7. Query comments from database (limit + 1 for cursor detection)
        8. Build next cursor if more data exists

        ### Query Parameters
        - `limit` — number of items per page (optional)
        - `cursor` — base64-encoded cursor for next page (optional)

        ### Error Cases
        - Invalid postId → `400` with "Invalid post id"
        - Limit < 0 → `400`
        - Limit > max → `400`
        - Not a member → `400` with "You are not a member of this server"

        ### Flow
        ```
        Request → Auth Middleware → Parse PostId → Validate Limit → Check Membership (DB) → Decode Cursor → Query Comments (DB) → Build Next Cursor → Response
        ```
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Post UUID
        in: path
        name: postId
        required: true
        type: string
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ServerCommentListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ValidationError'
      summary: Get comments on a post
      tags:
      - post-interactions
    post:
      consumes:
      - application/json
      description: |
        ## Create Comment Flow

        ### Overview
        Creates a comment on a post. Requires membership in the server where the post belongs. Supports nested replies via `parentId`.

        ### Auth
        Requires `Authorization` header with Bearer JWT access token.

        ### Business Logic
        1. Get `userId` from context
        2. Parse `postId` from URL path — must be a valid UUID
        3. Validate `content` (required, 1–1000 characters)
        4. If `parentId` provided, parse as UUID
        5. Check if user is a member of the server where the post belongs
        6. If `parentId` provided, verify parent comment exists and belongs to the same post
        7. Create comment record in database
        8. Return created comment details

        ### Error Cases
        - Invalid postId → `400` with "Invalid post id"
        - Content empty → `400` with "Content is required"
        - Content > 1000 chars → `400` with "Content must be at most 1000 characters"
        - Invalid parentId format → `400` with "Invalid parent comment id"
        - Not a member → `400` with "You are not a member of this server"
        - Parent comment not found → `400` with "Parent comment not found"

        ### Flow
        ```
        Request → Auth Middleware → Parse PostId → Validate Content → Parse ParentId → Check Membership (DB) → Check Parent Exists (DB) → Create Comment (DB) → Response
        ```
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Post UUID
        in: path
        name: postId
        required: true
        type: string
      - description: Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.ServerCommentCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ServerCommentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ValidationError'
      summary: Create a comment on a post
      tags:
      - post-interactions
  /posts/{postId}/comments/{commentId}:
    delete:
      description: |
        ## Delete Comment Flow

        ### Overview
        Deletes a comment from a post. Only the comment author can delete their own comment. Requires membership in the server where the post belongs.

        ### Auth
        Requires `Authorization` header with Bearer JWT access token.

        ### Business Logic
        1. Get `userId` from context
        2. Parse `postId` and `commentId` from URL path — must be valid UUIDs
        3. Check if user is a member of the server where the post belongs
        4. Check if user is the author of the comment
        5. Delete comment from database

        ### Error Cases
        - Invalid postId → `400` with "Invalid post id"
        - Invalid commentId → `400` with "Invalid comment id"
        - Not a member → `400` with "You are not a member of this server"
        - Not the author → `400` with "You are not the author of this comment"

        ### Flow
        ```
        Request → Auth Middleware → Parse Params → Check Membership (DB) → Check Comment Ownership (DB) → Delete Comment (DB) → Response (no data)
        ```
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Post UUID
        in: path
        name: postId
        required: true
        type: string
      - description: Comment UUID
        in: path
        name: commentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ValidationError'
      summary: Delete a comment
      tags:
      - post-interactions
  /posts/{postId}/likes:
    delete:
      description: |
        ## Unlike Post Flow

        ### Overview
        Removes a like from a post. Requires membership in the server where the post belongs. User must have previously liked the post.

        ### Auth
        Requires `Authorization` header with Bearer JWT access token.

        ### Business Logic
        1. Get `userId` from context
        2. Parse `postId` from URL path — must be a valid UUID
        3. Check if user is a member of the server where the post belongs
        4. Check if user has liked this post — return error if not liked
        5. Delete like record from database
        6. Fetch updated post to get new like count
        7. Return updated like count

        ### Error Cases
        - Invalid postId → `400` with "Invalid post id"
        - Not a member → `400` with "You are not a member of this server"
        - Not liked yet → `400` with "You haven't liked this post yet"

        ### Flow
        ```
        Request → Auth Middleware → Parse PostId → Check Membership (DB) → Check Already Liked (DB) → Delete Like (DB) → Get Updated Count (DB) → Response
        ```
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Post UUID
        in: path
        name: postId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PostLikeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ValidationError'
      summary: Unlike a post
      tags:
      - post-interactions
    post:
      description: |
        ## Like Post Flow

        ### Overview
        Likes a post. Requires membership in the server where the post belongs. Each user can only like a post once.

        ### Auth
        Requires `Authorization` header with Bearer JWT access token.

        ### Business Logic
        1. Get `userId` from context
        2. Parse `postId` from URL path — must be a valid UUID
        3. Check if user is a member of the server where the post belongs
        4. Check if user already liked this post — return error if already liked
        5. Create like record in database
        6. Fetch updated post to get new like count
        7. Return updated like count

        ### Error Cases
        - Invalid postId → `400` with "Invalid post id"
        - Not a member → `400` with "You are not a member of this server"
        - Already liked → `400` with "You already liked this post"

        ### Flow
        ```
        Request → Auth Middleware → Parse PostId → Check Membership (DB) → Check Not Already Liked (DB) → Create Like (DB) → Get Updated Count (DB) → Response
        ```
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Post UUID
        in: path
        name: postId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PostLikeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ValidationError'
      summary: Like a post
      tags:
      - post-interactions
...
  /servers/{serverId}/posts:
    get:
      description: |
        ## Get Server Posts Flow

        ### Overview
        Retrieves a paginated list of posts from a specific server. Requires server membership. Uses cursor-based pagination.

        ### Auth
        Requires `Authorization` header with Bearer JWT access token.

        ### Business Logic
        1. Get `userId` from context
        2. Parse `serverId` from URL path — must be a valid UUID
        3. Parse query parameters: `limit`, `cursor`
        4. Validate `limit` (>= 0 and <= max limit)
        5. Check if user is a member of the server
        6. If cursor provided, decode and unmarshal
        7. Query posts from database (limit + 1 for cursor detection) with author info and image URLs
        8. Build next cursor if more data exists

        ### Query Parameters
        - `limit` — number of items per page (optional)
        - `cursor` — base64-encoded cursor for next page (optional)

        ### Error Cases
        - Invalid serverId → `400` with "Invalid server id"
        - Limit < 0 → `400` with "Limit must be greater or equal than 0"
        - Limit > max → `400` with "Limit is exceeded max limit: {max}"
        - Not a member → `400` with "You are not a member of this server"

        ### Flow
        ```
        Request → Auth Middleware → Parse Params → Validate Limit → Check Membership (DB) → Decode Cursor → Query Posts (DB) → Build Next Cursor → Response
        ```
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Server UUID
        in: path
        name: serverId
        required: true
        type: string
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ServerPostListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ValidationError'
      summary: Get posts from a server
      tags:
      - posts
    post:
      consumes:
      - multipart/form-data
      description: |
        ## Create Post Flow

        ### Overview
        Creates a new post in a server. Requires server membership. Posts must include an image (uploaded as multipart form) and a caption.

        ### Auth
        Requires `Authorization` header with Bearer JWT access token.

        ### Business Logic
        1. Get `userId` from context
        2. Parse `serverId` from URL path — must be a valid UUID
        3. Check if user is a member of the server
        4. Read `image` file from multipart form (required, cannot be empty)
        5. Validate image file (format, size)
        6. Read `caption` from form value (required)
        7. Begin database transaction
        8. Upload image to MinIO (`server/post/{imageId}.webp`)
        9. Create post image record in database
        10. Create post record in database
        11. Commit transaction
        12. Fetch and return full post details including author info and image URL

        ### Error Cases
        - Invalid serverId → `400` with "Invalid server id"
        - Not a member → `400` with "You are not a member of this server"
        - No image file → `400` with "Image is required"
        - Invalid image format → `400` validation error
        - Caption empty → `400` with "Caption is required"

        ### Flow
        ```
        Request → Auth Middleware → Parse ServerId → Check Membership (DB) → Read Image → Validate Image → Read Caption → Begin TX → Upload Image (MinIO) → Create Image Record (DB) → Create Post (DB) → Commit TX → Get Post (DB) → Response
        ```
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Server UUID
        in: path
        name: serverId
        required: true
        type: string
      - description: Post image file
        in: formData
        name: image
        required: true
        type: file
      - description: Post caption
        in: formData
        name: caption
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ServerPostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ValidationError'
      summary: Create a new post
      tags:
      - posts
...
swagger: "2.0"